pyproject.toml
[project]
name = "app-strategist"
version = "0.1.0"
description = "Job search assistance CLI - compare resume/cover letter against job descriptions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "anthropic>=0.39.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.scripts]
app-strategist = "app_strategist.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app_strategist"]

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
